fix: resolve WebSocket test failures and implement group broadcasting functionality
- Fix WebSocket connection paths in retrospective tests to use retrospective_id instead of user_id
- Implement proper group broadcasting in RetrospectiveConsumer with channel_layer.group_send()
- Add message handlers for group_broadcast, retrospective_completed, insight_generated, and kpi_update
- Configure InMemoryChannelLayer for WebSocket tests to avoid Redis dependency
- Handle connection_established messages in test assertions
- Add asset/tests to pytest.ini testpaths for broader test discovery
- Fix permission logic in retrospective views for approval workflows

Resolves WebSocket timeout errors and enables real-time team collaboration features.
