+++ backend/retrospective/tests/test_tasks.py
@@ -449,4 +449,38 @@ class TaskIntegrationTest(TestCase):
             retrospective_id=str(retrospective.id),
             user_id=str(self.user.id)
         )
-        self.assertTrue(insight_result['success']) 
+        self.assertTrue(insight_result['success'])
+
+    @patch('retrospective.tasks.generate_retrospective.delay')
+    def test_celery_task_integration(self, mock_task):
+        """Test integration with Celery tasks"""
+        retrospective = RetrospectiveTask.objects.create(
+            campaign=self.campaign,
+            created_by=self.user
+        )
+        
+        # Mock Celery task
+        mock_task.return_value = MagicMock(id='test-task-id')
+        
+        # Start analysis (this would normally trigger Celery task)
+        from retrospective.views import RetrospectiveTaskViewSet
+        viewset = RetrospectiveTaskViewSet()
+        viewset.request = MagicMock(user=self.user)
+        
+        # This would normally call the Celery task
+        # In real implementation, this would be tested with Celery test worker
+        self.assertTrue(True)  # Placeholder for actual Celery integration test
+
+    def test_task_error_handling_and_recovery(self):
+        """Test task error handling and recovery mechanisms"""
+        import uuid
+        
+        # Test task with invalid retrospective ID
+        result = generate_retrospective('invalid-id')
+        self.assertFalse(result['success'])
+        self.assertIn('error', result)
+        
+        # Test task with non-existent retrospective
+        result = generate_retrospective(str(uuid.uuid4()))
+        self.assertFalse(result['success'])
+        self.assertIn('error', result)
