+++ docker-compose.yml
@@ -64,7 +64,6 @@ services:
       - DB_HOST=${DB_HOST:-host.docker.internal}
       - CELERY_BROKER_URL=${CELERY_BROKER_URL:-redis://redis:6379/0}
       - CELERY_RESULT_BACKEND=${CELERY_RESULT_BACKEND:-redis://redis:6379/0}
-      - DJANGO_SETTINGS_MODULE=backend.settings
     extra_hosts:
       - "host.docker.internal:host-gateway"
     depends_on:
